scripts/requirements.txt
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-core==1.29.5